% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{sass_compile}
\alias{sass_compile}
\title{Compile Sass strings or files}
\usage{
sass_compile(string, file, options = NULL)
}
\arguments{
\item{string}{Character vector of Sass strings to compile}

\item{file}{Filename of a Sass file}

\item{options}{A list of output options.}
}
\value{
A character vector of CSS.
}
\description{
Compile Sass strings or files
}
\examples{
sass_string <- "foo { margin: 10px * 2.5; }"
sass_compile(string = sass_string)

sass_string <- c(
  sass_string,
  "$bg-col:  blue; bar { background-color: $bg-col; }",
  "baz { color: lighten(#ff0000, 30\%); }"
)
sass_compile(string = sass_string)

test_file <- tempfile(fileext = "scss")
writeLines(sass_string, test_file)
sass_compile(file = test_file)
}
\seealso{
sass_compile_file sass_compile_string
}
