% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sass_compile_file}
\alias{sass_compile_file}
\title{Compile Sass file into CSS}
\usage{
sass_compile_file(filename, options = NULL)
}
\arguments{
\item{filename}{A Sass file (.scss, .sass)}

\item{options}{A list containing options for libsass}
}
\description{
This function compiles a file containing Sass code.
}
