% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sass_compile_string}
\alias{sass_compile_string}
\title{Compile Character Vector to CSS}
\usage{
sass_compile_string(strings, options = NULL)
}
\arguments{
\item{strings}{A character vector of Sass}

\item{options}{A list containing options for libsass}
}
\description{
This function compiles character vectors containing Sass code.
}
